regular expressions were grabbing too much
